service/requirements.txt
SQLAlchemy==1.3.16
numpy==1.18.5
opencv-python==4.2.0.34
psycopg2-binary==2.8.5
gevent==1.4.0
sqlalchemy-utils==0.36.8
Werkzeug==1.0.1
Flask==1.1.2
scipy==1.4.1
Flask-SocketIO==4.3.0
requests==2.23.0
redis==3.5.3
#grpcio-tools==1.0.0
grpcio==1.20.0
protobuf==3.13.0
environs==8.0.0